+++ env-proto.sh
@@ -4,7 +4,7 @@ set -euo pipefail
 BIN_DIR=${BIN_DIR:-/usr/local/bin}
 
 # renovate: datasource=github-releases depName=bufbuild/buf
-BUF_VERSION='v1.6.0'
+BUF_VERSION='v1.7.0'
 
 # Substitute BINARY_NAME for "buf", "protoc-gen-buf-breaking", or "protoc-gen-buf-lint".
 BINARY_NAME="buf"