+++ pkg/debuginfo/metadata_test.go
@@ -22,9 +22,10 @@ import (
 	"testing"
 
 	"github.com/go-kit/log"
+	"github.com/prometheus/client_golang/prometheus"
 	"github.com/stretchr/testify/require"
 	"github.com/thanos-io/objstore/client"
-	"github.com/thanos-io/objstore/filesystem"
+	"github.com/thanos-io/objstore/providers/filesystem"
 	"gopkg.in/yaml.v2"
 )
 
@@ -46,22 +47,12 @@ func TestMetadata(t *testing.T) {
 	})
 	require.NoError(t, err)
 
-	bucket, err := client.NewBucket(logger, cfg, "parca/store")
-	require.NoError(t, err)
-
-	cache, err := NewCache(
-		&CacheConfig{
-			Type: FILESYSTEM,
-			Config: &FilesystemCacheConfig{
-				Directory: cacheDir,
-			},
-		},
-	)
+	bucket, err := client.NewBucket(logger, cfg, prometheus.NewRegistry(), "parca/store")
 	require.NoError(t, err)
 
 	store, err := NewStore(
 		logger,
-		cache.Directory,
+		cacheDir,
 		NewObjectStoreMetadata(logger, bucket),
 		bucket,
 		NopDebugInfodClient{},