+++ pkg/query/query_test.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/apache/arrow/go/v8/arrow/memory"
 	"github.com/go-kit/log"
+	"github.com/polarsignals/frostdb"
 	columnstore "github.com/polarsignals/frostdb"
 	"github.com/polarsignals/frostdb/query"
 	"github.com/prometheus/client_golang/prometheus"
@@ -44,19 +45,20 @@ func Benchmark_Query_Merge(b *testing.B) {
 			logger := log.NewNopLogger()
 			reg := prometheus.NewRegistry()
 			tracer := trace.NewNoopTracerProvider().Tracer("")
-			col := columnstore.New(
+			col, err := columnstore.New(
+				logger,
 				reg,
-				8196,
-				64*1024*1024,
 			)
+			require.NoError(b, err)
 			colDB, err := col.DB("parca")
 			require.NoError(b, err)
+
+			schema, err := parcacol.Schema()
+			require.NoError(b, err)
+
 			table, err := colDB.Table(
 				"stacktraces",
-				columnstore.NewTableConfig(
-					parcacol.Schema(),
-				),
-				logger,
+				columnstore.NewTableConfig(schema),
 			)
 			require.NoError(b, err)
 			m := metastore.NewInProcessClient(metastoretest.NewTestMetastore(
@@ -76,7 +78,7 @@ func Benchmark_Query_Merge(b *testing.B) {
 			}
 
 			normalizer := parcacol.NewNormalizer(m)
-			ingester := parcacol.NewIngester(logger, normalizer, table)
+			ingester := parcacol.NewIngester(logger, normalizer, table, schema)
 
 			profiles, err := normalizer.NormalizePprof(ctx, "memory", p, false)
 			require.NoError(b, err)
@@ -119,5 +121,59 @@ func Benchmark_Query_Merge(b *testing.B) {
 				require.NoError(b, err)
 			}
 		})
+	}
+}
+
+func Benchmark_ProfileTypes(b *testing.B) {
+	// This benchmark is skipped by default as it requires a write-ahead log to
+	// be present for the "stacktraces" table in the "parca" database in
+	// "../../data".
+	b.Skip()
+
+	ctx := context.Background()
+	logger := log.NewNopLogger()
+	reg := prometheus.NewRegistry()
+	tracer := trace.NewNoopTracerProvider().Tracer("")
+	col, err := columnstore.New(
+		logger,
+		reg,
+		frostdb.WithWAL(),
+		frostdb.WithStoragePath("../../data"),
+	)
+	require.NoError(b, err)
+
+	require.NoError(b, col.ReplayWALs(ctx))
+
+	colDB, err := col.DB("parca")
+	require.NoError(b, err)
+
+	table, err := colDB.GetTable("stacktraces")
+	require.NoError(b, err)
+	table.Sync()
+
+	require.NoError(b, err)
+	m := metastore.NewInProcessClient(metastoretest.NewTestMetastore(
+		b,
+		logger,
+		reg,
+		tracer,
+	))
+
+	api := NewColumnQueryAPI(
+		logger,
+		tracer,
+		m,
+		getShareServerConn(b),
+		query.NewEngine(
+			memory.NewGoAllocator(),
+			colDB.TableProvider(),
+		),
+		"stacktraces",
+	)
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		_, err = api.ProfileTypes(ctx, &pb.ProfileTypesRequest{})
+		require.NoError(b, err)
 	}
 }