+++ pkg/query/top_test.go
@@ -50,7 +50,8 @@ func TestGenerateTopTable(t *testing.T) {
 	profiles, err := normalizer.NormalizePprof(ctx, "memory", p, false)
 	require.NoError(t, err)
 
-	symbolizedProfile, err := parcacol.SymbolizeNormalizedProfile(ctx, metastore, profiles[0])
+	tracer := trace.NewNoopTracerProvider().Tracer("")
+	symbolizedProfile, err := parcacol.NewArrowToProfileConverter(tracer, metastore).SymbolizeNormalizedProfile(ctx, profiles[0])
 	require.NoError(t, err)
 
 	res, err := GenerateTopTable(ctx, symbolizedProfile)
@@ -137,7 +138,7 @@ func TestGenerateTopTableAggregateFlat(t *testing.T) {
 	st2 := sres.Stacktraces[1]
 	st3 := sres.Stacktraces[2]
 
-	p, err := parcacol.SymbolizeNormalizedProfile(ctx, metastore, &profile.NormalizedProfile{
+	p, err := parcacol.NewArrowToProfileConverter(tracer, metastore).SymbolizeNormalizedProfile(ctx, &profile.NormalizedProfile{
 		Samples: []*profile.NormalizedSample{{
 			StacktraceID: st1.Id,
 			Value:        1,
@@ -204,7 +205,8 @@ func TestGenerateDiffTopTable(t *testing.T) {
 	}
 	require.Truef(t, found, "expected to find the specific sample")
 
-	p, err := parcacol.SymbolizeNormalizedProfile(ctx, metastore, profiles[0])
+	tracer := trace.NewNoopTracerProvider().Tracer("")
+	p, err := parcacol.NewArrowToProfileConverter(tracer, metastore).SymbolizeNormalizedProfile(ctx, profiles[0])
 	require.NoError(t, err)
 
 	res, err := GenerateTopTable(ctx, p)